Broke down template forums, created forums.py
