+++ app/templates/forums/forums.html
@@ -1,0 +1,15 @@
+{% extends 'forums/forums_main.html' %}
+
+{% block forums_create_post_button %} 
+  <a href="/forums/{{thread}}/create_post">
+    <button class="btn btn-primary btn-md mb-3">Create Post</button>
+  </a>
+{% endblock %}
+  
+{% block forums_post_titles %}
+  <div class="forum_post_titles"> 
+    {% include "forums/forums_post_titles.html" %}
+  </div>
+{% endblock %}  
+
+
