+++ app/templates/forums/forums_create_post.html
@@ -1,0 +1,25 @@
+{% extends '/forums/forums_main.html' %}
+
+    {% block forums_create_post_form %}
+        <!--Create Forum Post Form-->
+        {% if current_user.is_authenticated %}
+        <form class="forumForm mb-3" method="POST" action="/forums/create_post">
+            <!--Inherited forum id-->  
+            <input type="hidden" name="thread" value="{{thread}}">
+            <!--Title input-->
+            <div class="form-group">
+                <label>Title</label>
+                <input class="form-control" type="title" name="title" placeholder="Title" autofocus="">
+            </div>
+            <!--Body textarea input-->
+            <div class="form-group">
+                <label>Content</label>
+                <textarea class="form-control" rows="5" type="content" name="content" placeholder="Enter your text here"></textarea>
+            </div>
+            <!--Submit button-->
+            <button class="btn btn-primary btn-md" type="submit">Post</button>
+        </form>
+        {% else %}
+        Please sign in to create a post.
+        {% endif %}
+    {% endblock %}
