+++ app/templates/forums/forums_main.html
@@ -1,0 +1,10 @@
+{% extends "oneDiv.html" %}
+{% block oneDivContent %}
+
+    {% block forums_create_post_button %} {% endblock %}
+
+    {% block forums_create_post_form %} {% endblock %}
+ 
+    {% block forums_post_titles %} {% endblock %}  
+
+{% endblock %}
