+++ app/templates/forums/macros.html
@@ -1,0 +1,13 @@
+{# Forum post title view #}
+
+{% macro forum_post_title(post) %}
+  <article class="media content-section">
+    <div class="media-body">
+      <div class="article-metadata">
+        <a class="mr-2" href="#">{{ post.author }}</a>
+        <small class="text-muted">{{ post.date_posted }}</small>
+      </div>
+      <h2><a class="article-title" href="#">{{ post.title }}</a></h2>
+    </div>
+  </article>
+{% endmacro %}
